clean up tracing subscriber (no real span support implimented)
