+++ launcher/Application.cpp
@@ -159,7 +159,7 @@
 #include "WindowsConsole.h"
 #endif
 
-#include "hematite_static/src/debug.cxx.h"
+#include "hematite_static/src/log.cxx.h"
 
 #define STRINGIFY(x) #x
 #define TOSTRING(x) STRINGIFY(x)
