+++ launcher/hematite/include/log.h
