+++ launcher/hematite/src/log.rs
@@ -3,7 +3,7 @@
 pub mod ffi {
     #[namespace = "prism::hematite::log"]
     unsafe extern "C++" {
-        include!("hematite_static/debug.h");
+        include!("hematite-static/log.h");
         fn debug(msg: &str);
         fn warn(msg: &str);
         fn info(msg: &str);
@@ -17,6 +17,7 @@ pub mod ffi {
 fn setup_rust_tracing_qdebug() {
     use tracing_subscriber::prelude::*;
     tracing_subscriber::registry().with(QDebugLayer).init();
+    tracing::info!( rust_enabled = true, "rust tracing qDebug subscriber registered", );
 }
 
 #[derive(Default)]
@@ -37,7 +38,11 @@ impl tracing::field::Visit for QDebugVisitor {
         value: &(dyn std::error::Error + 'static),
     ) {
         self.field_lines
-            .push(format!("  field={} value={}", field.name(), value));
+            .push(format!("  field={} value=\"{}\"", field.name(), value));
+    }
+    fn record_str(&mut self, field: &tracing::field::Field, value: &str) {
+        self.field_lines
+            .push(format!("  field={} value={:?}", field.name(), value));
     }
     fn record_debug(&mut self, field: &tracing::field::Field, value: &dyn std::fmt::Debug) {
         self.field_lines
@@ -55,13 +60,7 @@ where
         event: &tracing::Event<'_>,
         _ctx: tracing_subscriber::layer::Context<'_, S>,
     ) {
-        let mut msg = indoc::formatdoc!(
-            "
-            [{:?}] {:?}
-                {:?}
-                
-            
-        ",
+        let mut msg = format!("  [{}] {} {}",
             event.metadata().level(),
             event.metadata().target(),
             event.metadata().name()
@@ -70,14 +69,14 @@ where
         let mut visitor = QDebugVisitor::default();
         event.record(&mut visitor);
         let field_lines: String = visitor.lines();
-        if field_lines.len() > 0 {
+        if !field_lines.is_empty() {
             msg.push_str("\n  ");
             msg.push_str(&field_lines);
         }
 
-        match event.metadata().level() {
-            &tracing::Level::WARN => ffi::warn(&msg),
-            &tracing::Level::INFO => ffi::info(&msg),
+        match *event.metadata().level() {
+            tracing::Level::WARN => ffi::warn(&msg),
+            tracing::Level::INFO => ffi::info(&msg),
             _ => ffi::debug(&msg),
         }
     }
